improve runUfsTests
